split components for search box
