+++ frontend/src/components/Atoms/SearchButton.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const SearchButton = styled.input`
+  cursor: pointer;
+  font-family: FontAwesome;
+  border: none;
+  background: #3879d9;
+  color: #fff;
+  position: absolute;
+  width: 4em;
+  height: 3.6em;
+  right: 0px;
+  top: -5px;
+  outline: none;
+`;
