+++ frontend/src/components/Atoms/SearchForm.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const SearchForm = styled.input`
+  border: none;
+  height: 1.8em;
+  width: 230px;
+  :focus {
+    outline: 0;
+  }
+  background-color: inherit;
+`;
