+++ frontend/src/components/Pages/PapersPage.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import { PaperSearchConditionType, PaperType } from "../../types";
 import Papers from "../Organisms/Papers";
-import SearchBox from "../Search/SearchBox";
-import SearchModal from "../Search/SearchModal";
+import SearchBox from "../Molecules/SearchBox";
+import SearchModal from "../Organisms/SearchModal";
 import MainButton from "../Atoms/MainButton";
 import { FetchPaperAPIService } from "../../api";
